Add --strip-chars and --strip-chars-additional
That enables user to specify what characters to strip, which is a good
thing.

By the way, just a remark: I do not consider --strip and accompanying
options to be braking the UNIX principles. One can argue that the same
could have been achieved with sed or awk, but in fact it's not possible,
because those tools can't distinguish between bookmarked text and
comment.